(PC-4932) Move up support.services to ath directory
